Add files via upload
tree.py defines a tree for the file system. api.py could be used for testing tree.py with functions required: ls, cd, mkdir, rm, etc.
server.py defines two class: DFSserver and DFSclient. They could connect to each other via a proper port. The tree defined in tree.py is imported in the file. However, it does not require api.py. Functions are re-write in the class DFSserver to suit its format.

Finished: tree. definition of all functions. functions of the server. 
To be done: functions and connection of the clients. connections need to be tested.